separated out the connection code, in addition to making a mess hehe
