+++ src/index.js
@@ -2,8 +2,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Chart from './components/Chart.jsx';
-import ConnectionManager from './ConnectionManager';
-import StreamManager from './StreamManager';
+import ConnectionManager from '../app/connection/ConnectionManager';
+import StreamManager from '../app/connection/StreamManager';
 import {TradeEndLine, TradeStartLine} from './components/VerticalLine.jsx';
 import MainStore from './store';
 import {MobxProvider} from './store/Connect';
@@ -17,21 +17,10 @@ class SmartChart extends React.Component {
     get connectionManager() { return this.chart.connectionManager; }
 
     async componentDidMount() {
-        let activeSymbols = this.props && this.props.activeSymbols;
-        if(!activeSymbols) {
-            const data = await this.connectionManager.send({
-                active_symbols: 'brief',
-                product_type: 'basic',
-            });
-            activeSymbols = data.active_symbols;
-        }
         this.mainStore.chart.isMobile = this.props.isMobile || false;
         this.mainStore.chart.onSymbolChange = this.props.onSymbolChange;
     }
-    componentWillReceiveProps({activeSymbols, onSymbolChange}) {
-        if(activeSymbols && activeSymbols !== this.chart.activeSymbols) {
-            this.chart.setActiveSymbols(activeSymbols);
-        }
+    componentWillReceiveProps({onSymbolChange}) {
         if(onSymbolChange && this.mainStore.chart.onSymbolChange !== onSymbolChange) {
             this.mainStore.chart.onSymbolChange = onSymbolChange;
         }
